requirements.txt
apache-beam==2.61.0
apache-flink==2.1.0
apache-flink-libraries==2.1.0
async-timeout==5.0.1
attrs==25.3.0
Automat==20.2.0
avro==1.12.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.14.2
blinker==1.9.0
cassandra-driver==3.29.3
cassandra-pylib==0.0.0
certifi==2025.10.5
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.3
click==8.3.0
cloud-init==25.2
cloudpickle==2.2.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crcmod==1.7
cryptography==3.4.8
curl_cffi==0.13.0
dbus-python==1.2.18
dill==0.3.1.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.8.0
docopt==0.6.2
fastavro==1.12.0
fasteners==0.20
find_libpython==0.4.1
Flask==3.1.2
flask-cors==6.0.1
frozendict==2.4.6
geomet==1.1.0
grpcio==1.65.5
hdfs==2.7.3
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jsonpatch==1.32
jsonpickle==3.4.2
jsonpointer==2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kafka-python==2.2.15
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.3
more-itertools==8.10.0
multitasking==0.0.12
netifaces==0.11.0
numpy==2.1.3
oauthlib==3.2.0
objsize==0.7.1
orjson==3.11.3
packaging==25.0
pandas==2.3.2
peewee==3.18.2
pemja==0.5.3
platformdirs==4.5.0
proto-plus==1.26.1
protobuf==5.29.5
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.7
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.23
pycurl==7.44.1
pydot==1.4.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.10.1
pymongo==4.15.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==5.4.1
redis==5.3.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.8
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.66
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.24.0
